refactor: MenuGroup 패키지 분리
