+++ src/main/java/kitchenpos/application/MenuService.java
@@ -4,9 +4,9 @@ import java.util.List;
 import java.util.stream.Collectors;
 import kitchenpos.application.dto.MenuResponse;
 import kitchenpos.dao.MenuDao;
-import kitchenpos.dao.MenuGroupDao;
 import kitchenpos.domain.Menu;
 import kitchenpos.domain.MenuProduct;
+import kitchenpos.menugroup.domain.MenuGroupDao;
 import kitchenpos.product.domain.Product;
 import kitchenpos.product.domain.ProductDao;
 import org.springframework.stereotype.Service;
