+++ src/main/java/kitchenpos/menugroup/domain/MenuGroup.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.menugroup.domain;
 
 public class MenuGroup {
     private final Long id;
