+++ src/main/java/kitchenpos/menugroup/ui/MenuGroupRequest.java
@@ -1,6 +1,6 @@
-package kitchenpos.ui.dto;
+package kitchenpos.menugroup.ui;
 
-import kitchenpos.domain.MenuGroup;
+import kitchenpos.menugroup.domain.MenuGroup;
 
 public class MenuGroupRequest {
